@@ -1,12 +1,23 @@
-module github.com/kevinanthony/gorps
+module github.com/kevinanthony/gorps/v2
 
-go 1.16
+go 1.21
 
 require (
-	github.com/go-chi/chi v1.5.4
-	github.com/json-iterator/go v1.1.11
+	github.com/go-chi/chi/v5 v5.0.12
+	github.com/json-iterator/go v1.1.12
 	github.com/pkg/errors v0.9.1
-	github.com/smartystreets/goconvey v1.6.4
-	github.com/stretchr/testify v1.3.0
-	goji.io v2.0.2+incompatible
+	github.com/smartystreets/goconvey v1.8.1
+	github.com/stretchr/testify v1.8.4
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gopherjs/gopherjs v1.17.2 // indirect
+	github.com/jtolds/gls v4.20.0+incompatible // indirect
+	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/smarty/assertions v1.15.0 // indirect
+	github.com/stretchr/objx v0.5.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
