@@ -5,19 +5,19 @@ import (
 )
 
 type testStruct struct {
-	Map    map[string]int `json:"map" xml:"-"`
-	Slice  []string       `json:"slice" xml:"slice"`
+	Map    map[string]int `json:"map"    xml:"-"`
+	Slice  []string       `json:"slice"  xml:"slice"`
 	String string         `json:"string" xml:"string"`
-	Int    int            `json:"int" xml:"int"`
-	Float  float32        `json:"float" xml:"flat"`
-	Time   time.Time      `json:"time" xml:"time"`
-	Struct subStrict      `json:"sub" xml:"struct"`
+	Int    int            `json:"int"    xml:"int"`
+	Float  float32        `json:"float"  xml:"flat"`
+	Time   time.Time      `json:"time"   xml:"time"`
+	Struct subStrict      `json:"sub"    xml:"struct"`
 }
 
 type subStrict struct {
 	String string  `json:"string" xml:"string"`
-	Int    int     `json:"int" xml:"int"`
-	Float  float32 `json:"float" xml:"flat"`
+	Int    int     `json:"int"    xml:"int"`
+	Float  float32 `json:"float"  xml:"flat"`
 }
 
 func newTestStruct() testStruct {
@@ -27,7 +27,7 @@ func newTestStruct() testStruct {
 		String: "something",
 		Int:    42,
 		Float:  3.1415,
-		Time:   time.Date(1989, 11, 9, 18, 01, 00, 00, time.UTC),
+		Time:   time.Date(1989, 11, 9, 18, 0o1, 0o0, 0o0, time.UTC),
 		Struct: subStrict{
 			String: "else",
 			Int:    99,
