@@ -1,49 +1,68 @@
+// Code generated by mockery v2.35.4. DO NOT EDIT.
+
 package http
 
 import (
-	"io"
-	"net/http"
+	nethttp "net/http"
 
-	"github.com/stretchr/testify/mock"
+	mock "github.com/stretchr/testify/mock"
 )
 
-var _ Client = (*ClientMock)(nil)
-var _ Native = (*NativeMock)(nil)
-var _ io.ReadCloser = (*BodyMock)(nil)
-
+// ClientMock is an autogenerated mock type for the Client type
 type ClientMock struct {
 	mock.Mock
 }
 
-func (c *ClientMock) Do(req *http.Request, v interface{}) error {
-	return c.Called(req, v).Error(0)
-}
+// Do provides a mock function with given fields: req
+func (_m *ClientMock) Do(req *nethttp.Request) ([]byte, error) {
+	ret := _m.Called(req)
 
-type NativeMock struct {
-	mock.Mock
-}
-
-func (n *NativeMock) Do(req *http.Request) (*http.Response, error) {
-	args := n.Called(req)
-
-	var resp *http.Response
-	if item := args.Get(0); item != nil {
-		resp = item.(*http.Response)
+	var r0 []byte
+	var r1 error
+	if rf, ok := ret.Get(0).(func(*nethttp.Request) ([]byte, error)); ok {
+		return rf(req)
+	}
+	if rf, ok := ret.Get(0).(func(*nethttp.Request) []byte); ok {
+		r0 = rf(req)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).([]byte)
+		}
 	}
 
-	return resp, args.Error(1)
+	if rf, ok := ret.Get(1).(func(*nethttp.Request) error); ok {
+		r1 = rf(req)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
 }
 
-type BodyMock struct {
-	mock.Mock
+// DoAndUnmarshal provides a mock function with given fields: req, v
+func (_m *ClientMock) DoAndUnmarshal(req *nethttp.Request, v interface{}) error {
+	ret := _m.Called(req, v)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(*nethttp.Request, interface{}) error); ok {
+		r0 = rf(req, v)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
 }
 
-func (b *BodyMock) Read(p []byte) (n int, err error) {
-	args := b.Called(p)
+// NewClientMock creates a new instance of ClientMock. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewClientMock(t interface {
+	mock.TestingT
+	Cleanup(func())
+}) *ClientMock {
+	mock := &ClientMock{}
+	mock.Mock.Test(t)
 
-	return args.Int(0), args.Error(1)
-}
+	t.Cleanup(func() { mock.AssertExpectations(t) })
 
-func (b *BodyMock) Close() error {
-	return b.Called().Error(0)
+	return mock
 }
