@@ -3,6 +3,7 @@ package http
 import (
 	"context"
 	"fmt"
+	"io"
 	native "net/http"
 	"net/url"
 	"strings"
@@ -10,30 +11,50 @@ import (
 	"github.com/pkg/errors"
 )
 
+//nolint:interfacebloat
+//go:generate mockery --name=RequestBroker --structname=RequestBrokerMock --filename=request_broker_mock.go --inpackage
 type RequestBroker interface {
-	Go(ctx context.Context, v interface{}) error
+	DoAndUnmarshal(ctx context.Context, v interface{}) error
+	Do(ctx context.Context) ([]byte, error)
 
 	Post() RequestBroker
 	Get() RequestBroker
 	Put() RequestBroker
 	Delete() RequestBroker
 
-	Domain(string) RequestBroker
-	Path(string) RequestBroker
-	Parameter(string, string) RequestBroker
+	URL(url string, v ...any) RequestBroker
+	Query(key string, value string) RequestBroker
+	Header(key string, value string) RequestBroker
+	Body(body string) RequestBroker
 
-	Header(string, string) RequestBroker
+	CreateRequest(ctx context.Context) (*native.Request, error)
 }
 
 type requestBroker struct {
 	err    error
 	client Client
-	domain string
-	path   string
-	method MethodType
+	url    string
 
-	parameters map[string]string
-	headers    map[string]string
+	method  MethodType
+	headers map[string]string
+	query   map[string]string
+	body    string
+}
+
+func NewRequest(client Client) RequestBroker {
+	r := &requestBroker{
+		method:  MethodGet,
+		headers: map[string]string{},
+		query:   map[string]string{},
+	}
+
+	if client == nil {
+		r.setErrStr("native client is nil")
+	}
+
+	r.client = client
+
+	return r
 }
 
 func (r *requestBroker) Post() RequestBroker {
@@ -60,20 +81,14 @@ func (r *requestBroker) Delete() RequestBroker {
 	return r
 }
 
-func (r *requestBroker) Domain(s string) RequestBroker {
-	r.domain = s
+func (r *requestBroker) URL(s string, v ...any) RequestBroker {
+	r.url = fmt.Sprintf(s, v...)
 
 	return r
 }
 
-func (r *requestBroker) Path(s string) RequestBroker {
-	r.path = s
-
-	return r
-}
-
-func (r *requestBroker) Parameter(pattern, value string) RequestBroker {
-	r.parameters[pattern] = value
+func (r *requestBroker) Query(pattern, value string) RequestBroker {
+	r.query[pattern] = value
 
 	return r
 }
@@ -84,58 +99,76 @@ func (r *requestBroker) Header(header, value string) RequestBroker {
 	return r
 }
 
-func NewRequest(client Client) RequestBroker {
-	r := &requestBroker{
-		method:     MethodGet,
-		parameters: map[string]string{},
-		headers:    map[string]string{},
-	}
-
-	if client == nil {
-		r.setErrStr("native client is nil")
-	}
-
-	r.client = client
+func (r *requestBroker) Body(bytes string) RequestBroker {
+	r.body = bytes
 
 	return r
 }
 
-func (r *requestBroker) Go(ctx context.Context, out interface{}) error {
-	if r.err != nil {
-		return r.err
+func (r *requestBroker) DoAndUnmarshal(ctx context.Context, out interface{}) error {
+	req, err := r.CreateRequest(ctx)
+	if err != nil {
+		return err
 	}
 
-	for k, v := range r.parameters {
-		if !strings.Contains(r.path, k) {
-			return fmt.Errorf("missing parameter %s in path", k)
-		}
+	return r.client.DoAndUnmarshal(req, out)
+}
 
-		r.path = strings.ReplaceAll(r.path, k, v)
+func (r *requestBroker) Do(ctx context.Context) ([]byte, error) {
+	req, err := r.CreateRequest(ctx)
+	if err != nil {
+		return nil, err
 	}
 
-	req := &native.Request{
-		Method: string(r.method),
-		URL: &url.URL{
-			Scheme: "https",
-			Host:   r.domain,
-			Path:   r.path,
-		},
-		Proto:  "https",
-		Header: native.Header{},
-		Body:   nil,
-	}
-
-	for k, v := range r.headers {
-		req.Header.Add(k, v)
-	}
-
-	req = req.WithContext(ctx)
-
-	return r.client.Do(req, out)
+	return r.client.Do(req)
 }
 
 func (r *requestBroker) setErrStr(s string) {
 	if r.err != nil {
 		r.err = errors.New(s)
 	}
+}
+
+func (r *requestBroker) CreateRequest(ctx context.Context) (*native.Request, error) {
+	if r.err != nil {
+		return nil, r.err
+	}
+
+	url, err := url.Parse(r.url)
+	if err != nil {
+		return nil, err
+	}
+
+	req := &native.Request{
+		Method:     string(r.method),
+		URL:        url,
+		Proto:      "HTTP/1.1",
+		ProtoMajor: 1,
+		ProtoMinor: 1,
+		Header:     make(native.Header),
+		Body:       nil,
+		Host:       url.Host,
+	}
+
+	if len(r.body) > 0 {
+		req.ContentLength = int64(len(r.body))
+		req.Body = io.NopCloser(strings.NewReader(r.body))
+		req.GetBody = func() (io.ReadCloser, error) {
+			return io.NopCloser(strings.NewReader(r.body)), nil
+		}
+	}
+
+	for k, v := range r.headers {
+		req.Header.Add(k, v)
+	}
+
+	query := req.URL.Query()
+
+	for k, v := range r.query {
+		query.Set(k, v)
+	}
+
+	req.URL.RawQuery = query.Encode()
+
+	return req.WithContext(ctx), nil
 }
